+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Documentation of release versions of `nacc-attribute-deriver`
     * Make `affiliate` an integer
     * Update `NACCREFR` to not allow I4 9s (unknowns) override previously known values
 * rc4:
+    * Adds new variables for BDS scope
     * Changes `INFORMED_BLANK` string to be `__blank__` to avoid collisions with actual write-in values
     * Update deceased status to not be unset if participant has an NP form (so death is certain)